Move billing details function to correct file
This has nothing to do with the ft_billing table.
