+++ app/platform_stats/rest.py
@@ -4,7 +4,6 @@ from flask import Blueprint, jsonify, request
 
 from app.dao.date_util import get_financial_year_for_datetime
 from app.dao.fact_billing_dao import (
-    fetch_billing_details_for_all_services,
     fetch_daily_sms_provider_volumes_for_platform,
     fetch_daily_volumes_for_platform,
     fetch_usage_for_all_services_letter,
@@ -15,6 +14,7 @@ from app.dao.fact_billing_dao import (
 from app.dao.fact_notification_status_dao import (
     fetch_notification_status_totals_for_all_services,
 )
+from app.dao.services_dao import fetch_billing_details_for_all_services
 from app.errors import InvalidRequest, register_errors
 from app.models import UK_POSTAGE_TYPES
 from app.platform_stats.platform_stats_schema import platform_stats_request
